fix: change function name to DesignPage in pages folder, and refactor back folder name
