+++ src/components/Design/Colors/Primary/PrimaryColor.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 
-import Colors from '../../Colors';
+import Colors from '..';
 
-const pimaryColor = {
+const primaryColor = {
   cyanDark: '#2B788B',
   cyanLight: '#C3DCE3',
   pinkDark: '#945069',
   pinkLight: '#F2D4DC',
-  Black: '#000000',
+  black: '#000000',
   greyLight: '#F6F5F4',
   greyMedium: '#E0E0E0',
   greyIcon: '#BABABA',
@@ -17,7 +17,7 @@ const pimaryColor = {
 const PrimaryColor = () => (
   <div>
     <h2>---Primary Color---</h2>
-    {Object.entries(pimaryColor).map(([colorName, colorValue]) => (
+    {Object.entries(primaryColor).map(([colorName, colorValue]) => (
       <Colors key={colorName} colorValue={colorValue}>
         {colorName.charAt(0).toUpperCase() + colorName.slice(1)}
       </Colors>
