--- src/components/DesignPage/Colors/Primary/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './PrimaryColor';
