--- src/components/DesignPage/Colors/Secondary/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './SecondaryColor';
