--- src/components/DesignPage/Colors/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Colors';
