--- src/components/DesignPage/DesignPage.module.scss
@@ -1,19 +1,0 @@
-.designContainer {
-  display: flex;
-  gap: 100px;
-}
-.fontBaseStyle {
-  font-family: 'Dela Gothic One';
-  font-style: normal;
-  letter-spacing: -0.04em;
-  line-height: 120%;
-  font-size: 32px;
-}
-
-.headingBaseStyle {
-  font-family: 'Dela Gothic One';
-  font-style: normal;
-  letter-spacing: -0.04em;
-  line-height: 120%;
-  color: #000000;
-}
