--- src/components/DesignPage/Headings/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Headings';
